:recycle:refactor:Remove Unused comments and Move directories
