+++ src/main/java/excluz/excluz/auth/util/SecurityContextUtil.java
@@ -3,10 +3,8 @@ package excluz.excluz.auth.util;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import excluz.excluz.common.exception.NotFoundException;
-import excluz.excluz.common.exception.error.ErrorCode;
+
 import excluz.excluz.domain.user.enums.UserRole;
-import lombok.extern.slf4j.Slf4j;
 
 public class SecurityContextUtil {
