+++ src/main/java/excluz/excluz/common/exception/error/ErrorResponseDto.java
@@ -2,7 +2,6 @@ package excluz.excluz.common.exception.error;
 
 import org.springframework.http.HttpStatus;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
