+++ src/main/java/excluz/excluz/domain/event/eventApplicant/repository/EventApplicantRepository.java
@@ -3,9 +3,7 @@ package excluz.excluz.domain.event.eventApplicant.repository;
 import excluz.excluz.common.entity.EventApplicant;
 import excluz.excluz.common.entity.Event;
 import excluz.excluz.domain.event.eventApplicant.enums.ApplicantStatus;
-import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
