+++ src/main/java/excluz/excluz/domain/event/eventItem/dto/EventItemDto.java
@@ -1,9 +1,6 @@
 package excluz.excluz.domain.event.eventItem.dto;
 
-import excluz.excluz.common.entity.Event;
 import excluz.excluz.common.entity.EventItem;
-import excluz.excluz.common.entity.Item;
-import jakarta.validation.constraints.Min;
 import lombok.Builder;
 import lombok.Getter;
