+++ src/main/java/excluz/excluz/domain/event/eventItem/dto/EventItemRequestDto.java
@@ -3,7 +3,6 @@ package excluz.excluz.domain.event.eventItem.dto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
