+++ src/main/java/excluz/excluz/domain/order/order/controller/OrderController.java
@@ -12,12 +12,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.web.bind.annotation.*;
 import excluz.excluz.auth.util.SecurityContextUtil;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
