+++ src/main/java/excluz/excluz/domain/store/item/repository/ItemRepository.java
@@ -1,6 +1,5 @@
 package excluz.excluz.domain.store.item.repository;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.domain.Page;
