+++ src/main/java/excluz/excluz/domain/user/repository/UserRepository.java
@@ -6,9 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import excluz.excluz.common.entity.Streamer;
 import excluz.excluz.common.entity.User;
-import jakarta.validation.constraints.NotBlank;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
